Create dice.rs
Implemented the Die and Roll data structures.
Adding an instance of Die to the Roll instance to more readily distinguish each Roll instance in a collection.
Using the readonly crate, marked Roll as being readonly so it can only be modified from within this module; Necessitated creation of a Roll::new() constructor method.
Implemented automated tests for this module.